move SocrataApi class to its own module
