+++ covid19_sfbayarea/data/marin.py
@@ -2,7 +2,7 @@ from typing import List, Dict, Iterable
 from datetime import datetime
 from ..errors import FormatError
 from ..utils import assert_equal_sets, parse_datetime
-from .utils import SocrataApi
+from .socrata import SocrataApi
 
 
 API_IDS = {
